Moved generateRandomName to MainActivity so it can read resource file
