+++ QRhunter/.idea/deploymentTargetDropDown.xml
@@ -12,6 +12,6 @@
         </deviceKey>
       </Target>
     </runningDeviceTargetSelectedWithDropDown>
-    <timeTargetWasSelectedWithDropDown value="2023-03-11T23:17:20.017677600Z" />
+    <timeTargetWasSelectedWithDropDown value="2023-03-11T23:21:38.181736700Z" />
   </component>
 </project>
